introduce RtnlBuffer and move some tests
It's more consistent to use XxxBuffer types everywhere, instead of
having a custom private "parse()" message for a given XxxMessage type.
